requirements.txt
numpy
opencv-python
loguru
scikit-image
scikit-learn
tqdm
torchvision>=0.10.0
Pillow
thop
ninja
tabulate
tensorboard
lap
motmetrics
filterpy
h5py
matplotlib
scipy
prettytable
easydict
tensorboard
pyyaml
yacs
termcolor
gdown
onnx
onnxruntime
onnx-simplifier


